Reload design when workspace configuration changes
If supported by the client, the server registers for changes to the
configuration file using client/registerCapability.
When the configuration file changes, the client send a notification
workspace/didChangeWatchedFiles to the server. The configuration is
then loaded again and project configuration is updated:
The design state is reset, new files are added and parsed. As required
for incremental document updates, source files that are already loaded
are parsed from in-memory sources.