Improved completions (#313)
This commit adds specific completions for selected names and attributes. It also extends the completion capabilities to entities, packages and package bodies (previously, this was only possible in architectures). Furthermore, all visible symbols from entities, packages and architectures are included for the generic completion provider. Previously, only symbols in the declarative part of architectures were includes.